requirements.txt
absl-py==0.7.1
appdirs==1.4.3
appnope==0.1.0
arrow==0.14.2
art==3.6
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
blis==0.2.4
Bottleneck==1.2.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
contextlib2==0.6.0.post1
coverage==4.5.3
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
Cython==0.29.12
decorator==4.4.0
defusedxml==0.6.0
dlib==19.17.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.14
easydict==1.9
editdistance==0.5.3
entrypoints==0.3
eventlet==0.25.0
face-recognition==1.2.3
face-recognition-models==0.3.0
fastai==1.0.55
fastprogress==0.1.21
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Ext==0.1
Flask-PyMongo==2.3.0
Flask-SocketIO==4.0.0
gast==0.2.2
gevent==1.4.0
gluoncv==0.5.0
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
imbalanced-learn==0.4.3
imblearn==0.0
importlib-metadata==0.23
imutils==0.5.2
inform==1.17.0
insightface==0.1.5
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jwt==0.6.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==19.2.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==7.2.0
mqttclient==0.0.2
murmurhash==1.0.2
mxnet==1.4.1
Naked==0.1.31
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numexpr==2.6.9
numpy==1.16.1
nvidia-ml-py3==7.352.0
olefile==0.46
onnx==1.5.0
onoff==1.0.1
opencv-python==4.1.0.25
packaging==19.1
paho-mqtt==1.4.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.5.0.1
plac==0.9.6
preshed==2.0.1
prometheus-client==0.7.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
PyAudio==0.2.11
pycm==2.2
pycparser==2.19
pycryptodome==3.8.2
Pygments==2.4.2
PyJWT==1.7.1
pymongo==3.8.0
pynvx==1.0.0
pyparsing==2.4.0
pypng==0.0.19
pyrsistent==0.15.2
pytesseract==0.3.0
python-dateutil==2.8.0
python-engineio==3.7.0
python-socketio==4.0.3
python-vlc==3.0.6109
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.5.1
quantiphy==2.5.0
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
rtsp==1.1.6
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.2
spacy==2.1.6
srsly==0.0.7
statistics==1.0.3.5
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-object-detection-api==0.1.1
termcolor==1.1.0
terminado==0.8.2
tesseract==0.1.3
testpath==0.4.2
thinc==7.0.8
tk==0.1.0
torch==1.1.0.post2
torchvision==0.3.0
tornado==6.0.2
tqdm==4.32.2
traitlets==4.3.2
twine==3.0.0
typing==3.6.6
typing-extensions==3.7.2
urllib3==1.25.3
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wheezy.routing==0.1.157
widgetsnbextension==3.4.2
zipp==0.6.0
